cold-fixes

+++ app/Repositories/RequestRepository.php
@@ -33,30 +33,33 @@ class RequestRepository implements RequestRepositoryInterface
     public function update($id, $data)
     {
         $review_request = ReviewRequest::findOrFail($id);
-
         $auth_user_id = Auth::user()->id;
-        $isReputationUp = ($data->reputation > $review_request->reputation);
-        $isReputationDown = ($data->reputation < $review_request->reputation);
-        $review_request->reputation = $data->reputation;
+
+        // Check if the reputation change and Up or Down
+        if ($data->has('reputation')) {
+            $isReputationUp =  ($data->reputation > $review_request->reputation);
+            $isReputationDown = ($data->reputation < $review_request->reputation);
+            $review_request->reputation = $data->reputation;
+
+            // If reputation change save user vote  or delete his vote
             if ($isReputationUp) {
+                $review_request->votes()->detach($auth_user_id); // temp solution, for removing dubl votes
                 $review_request->votes()->attach($auth_user_id);
             } elseif ($isReputationDown) {
                 $review_request->votes()->detach($auth_user_id);
             }
-        $review_request->save();
-        if ($review_request->user_id == $auth_user_id) {
-            $review_request->title = $data->title;
-            $review_request->details = $data->details;
-           
-
-            // There is may be another fields witch need to update
-
-            $review_request->save();
-
-            return $review_request;
-        } else {
-            return ['error' => ['message' => 'You can not edit not yours request']];
         }
+
+        // Fill only existing fields (see http://ryanchenkie.com/laravel-put-requests/)
+        if ($review_request->user_id == $auth_user_id) {
+            $review_request->title = $data->title ? $data->title : $review_request->title;
+            $review_request->details = $data->has('details') ? $data->details: $review_request->details;
+            // Another fields witch are need to update ...
+        }
+
+        $review_request->save();
+
+        return $review_request;
 
     }
